add: line and text_handler module
