+++ text_handler/__init__.py
@@ -1,0 +1,5 @@
+from .self_ref_exp import self_ref_exp
+
+__all__ = [
+    self_ref_exp,
+]
